src/pollutionService/pyproject.toml
[tool.poetry]
name = "pollutionservice"
version = "0.1.0"
description = ""
authors = ["Mostafa <mostafa.babaii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
python-dotenv = "^0.21.1"
celery = "^5.2.7"
flower = "^1.2.0"
pydantic = "^1.10.4"
elasticsearch = "7.9"
elasticsearch-dsl = "7.4.0"
djangorestframework = "^3.14.0"
pika = "^1.3.1"
watchdog = "^2.2.1"
gevent = "^22.10.2"
redis = "^4.4.2"
requests = "^2.28.2"
psycopg2-binary = "^2.9.5"
django-elasticsearch-dsl = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
